+++ src/main/java/model/polygon/Rectangle.java
@@ -1,4 +1,4 @@
-package model.coordinatecalculation.polygon;
+package model.polygon;
 
 import model.DistinguishedCoordinates;
 import model.Point;
